+++ x/staking/keeper/genesis.go
@@ -60,10 +60,7 @@ func (k Keeper) InitGenesis(ctx sdk.Context, data *types.GenesisState) (res []ab
 	}
 
 	for _, delegation := range data.Delegations {
-		delegatorAddress, err := sdk.AccAddressFromBech32(delegation.DelegatorAddress)
-		if err != nil {
-			panic(err)
-		}
+		delegatorAddress := sdk.MustAccAddressFromBech32(delegation.DelegatorAddress)
 
 		// Call the before-creation hook if not exported
 		if !data.Exported {